day3/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.28.1
python-dotenv==1.1.1
numpy==1.26.4
sentence-transformers>=2.7.0
